requirements.txt
beautifulsoup4
lxml
google-genai
python-dotenv
flask-flask-cors